# Remove redundant bounds check in mul! with sparse matrices